+++ CHANGELOG.md
@@ -20,6 +20,7 @@ v0.15.0 (?? ??? 2018)
 - Added support for Image text wrapping distance @troosan #1310
 - Added parsing of CSS line-height and text-indent in HTML reader @troosan #1316
 - Added the ability to enable gridlines and axislabels on charts @FrankMeyer #576
+- Add support for table indent (tblInd) @Trainmaster #1343
 - Added parsing of internal links in HTML reader @lalop #1336
 
 ### Fixed